requirements_dev.txt
df_engine>=0.8.1
requests>=2.26.0
streamlit>=1.1.0
graphviz==0.17
plotly==5.5.0
infi.clickhouse-orm==2.1.1
ipywidgets==7.6.5
traitlets==5.1.1
psycopg2==2.9.2
SQLAlchemy==1.4.27
sphinx>=1.7.9
sphinx_rtd_theme>=0.4.0
pytest
black